test: update testing ById func signature

+++ internal/certificate/testing.go
@@ -45,8 +45,8 @@ func Sweep(r string) error {
 }
 
 // ByID returns a function that obtains a certificate by its ID.
-func ByID(t *testing.T, cert *hcloud.Certificate) func(*hcloud.Client, int) bool {
-	return func(c *hcloud.Client, id int) bool {
+func ByID(t *testing.T, cert *hcloud.Certificate) func(*hcloud.Client, int64) bool {
+	return func(c *hcloud.Client, id int64) bool {
 		found, _, err := c.Certificate.GetByID(context.Background(), id)
 		if err != nil {
 			t.Fatalf("find certificate %d: %v", id, err)
