+++ internal/firewall/testing.go
@@ -45,8 +45,8 @@ func Sweep(r string) error {
 }
 
 // ByID returns a function that obtains a firewall by its ID.
-func ByID(t *testing.T, firewall *hcloud.Firewall) func(*hcloud.Client, int) bool {
-	return func(c *hcloud.Client, id int) bool {
+func ByID(t *testing.T, firewall *hcloud.Firewall) func(*hcloud.Client, int64) bool {
+	return func(c *hcloud.Client, id int64) bool {
 		found, _, err := c.Firewall.GetByID(context.Background(), id)
 		if err != nil {
 			t.Fatalf("find firewall %d: %v", id, err)
